+++ src/main/java/com/softeer/backend/fo_domain/draw/dto/history/DrawHistoryResponseDto.java
@@ -1,4 +1,4 @@
-package com.softeer.backend.fo_domain.draw.dto.result;
+package com.softeer.backend.fo_domain.draw.dto.history;
 
 import lombok.Data;
 import lombok.experimental.SuperBuilder;
