+++ src/main/java/com/softeer/backend/fo_domain/draw/dto/history/DrawHistoryWinnerResponseDto.java
@@ -1,4 +1,4 @@
-package com.softeer.backend.fo_domain.draw.dto.result;
+package com.softeer.backend.fo_domain.draw.dto.history;
 
 import com.softeer.backend.fo_domain.draw.dto.modal.WinModal;
 import lombok.Data;
